test(sybra): gate e2e tests behind build tag
Tag all e2e test files with //go:build e2e so go test ./...
runs only fast unit tests by default (~3s vs 30-60s).

- Add //go:build e2e to e2e_*.go, orchestrator_resume, svc_orchestrator,
  svc_reviews (all spawn real agent processes)
- Extract TestPRMonitorEligible → app_reviews_unit_test.go (no agents)
- Extract TestPRIssueKindConstants → workflow_enum_crosscheck_test.go
- Add BenchmarkStartWorkflow/AdvanceStep/Retry/ResumeStalled/Concurrent
  in internal/workflow/engine_bench_test.go (baselines for perf work)
- Add test-go-e2e CI job: go test -tags e2e ./internal/sybra/...
