+++ .github/workflows/ci.yml
@@ -24,6 +24,21 @@ jobs:
       - name: Run Go tests
         run: go test ./...
 
+  test-go-e2e:
+    name: Go E2E Tests
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+
+      - uses: jdx/mise-action@1648a7812b9aeae629881980618f079932869151 # v4.0.1
+
+      - name: Build frontend (needed for go:embed)
+        run: cd frontend && npm ci && npm run build:desktop
+
+      - name: Run Go e2e tests
+        run: go test -tags e2e -timeout 10m ./internal/sybra/...
+
   test-frontend:
     name: Frontend Tests
     runs-on: ubuntu-latest
