+++ internal/sybra/svc_reviews_test.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package sybra
 
 import (
